Experimenting with test-only eleventy config.
