+++ src/site/_data/housing.js
@@ -1,8 +1,7 @@
 const { AssetCache } = require("@11ty/eleventy-fetch");
 var Airtable = require('airtable');
-if (process.env.SITE_CONTEXT != 'test') {
-  var base = new Airtable({ apiKey: process.env.AIRTABLE_API_KEY }).base(process.env.AIRTABLE_BASE_ID);
-}
+var base = new Airtable({ apiKey: process.env.AIRTABLE_API_KEY }).base(process.env.AIRTABLE_BASE_ID);
+
 
 const UNITS_TABLE = "tblRtXBod9CC0mivK";
 const HOUSING_DATABASE_TABLE = "tbl8LUgXQoTYEw2Yh";
@@ -223,10 +222,6 @@ const filterOptions = (housing) => {
 // available in the Airtable data at that time.
 module.exports = async function() {
 
-  if(process.env.SITE_CONTEXT == 'test') {
-    return {};
-  }
-  
   const asset = new AssetCache("affordable_housing_data");
   // This cache duration will only be used at build time.
   let cacheDuration = "1h";
