+++ src/site/static/pages.11tydata.js
@@ -1,8 +1,6 @@
 const { AssetCache } = require("@11ty/eleventy-fetch");
 var Airtable = require('airtable');
-if (process.env.SITE_CONTEXT != 'test') {
-  var base = new Airtable({ apiKey: process.env.AIRTABLE_API_KEY }).base(process.env.AIRTABLE_BASE_ID);
-}
+var base = new Airtable({ apiKey: process.env.AIRTABLE_API_KEY }).base(process.env.AIRTABLE_BASE_ID);
 
 const fetchSection = (id) => {
   const table = base("tblAkC6dlPJc4o0Je"); // sections table
